requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
openai>=1.3.0
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.3
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
